Count trains used also in scoring
Till now, for scoring I only checked if the two points of a ticket
existed in the given route, and if yes, the score was increased
accordingly. But I think that the max score may go over 255, so I
guard for this. I'm not sure if the maximum score, now that I think of
it, may go over 255, but just to be on the safe side, I converted the
score to 16 bit unsigned, instead of 8 bit unsigned. If this is not
the case, I may bring it back to u8.

+++ src/scoring.rs
@@ -1,18 +1,18 @@
 use ticket2ride::{create_network, get_big_tickets, get_tickets, City};
 
-pub fn get_scores(route: Vec<City>) -> u8 {
-    let mut score: u8 = 0;
+pub fn get_scores(route: Vec<City>) -> u16 {
+    let mut score: u16 = 0;
     let routes = create_network();
     let big_tickets = get_big_tickets();
     let small_tickets = get_tickets();
     for ticket in big_tickets.iter() {
         if route.iter().any(|&i| i == ticket.depart) && route.iter().any(|&i| i == ticket.arrive) {
-            score += ticket.value;
+            score += ticket.value as u16;
         }
     }
     for ticket in small_tickets.iter() {
         if route.iter().any(|&i| i == ticket.depart) && route.iter().any(|&i| i == ticket.arrive) {
-            score += ticket.value;
+            score += ticket.value as u16;
         }
     }
     // Get score from used trains
